Improve testing & implement intersection - Implement intersection directly in index.js - Replaced Geometry Testing with tape test - Colorized geojson outputs - Update JSDocs offset
