+++ packages/turf-line-offset/test/in/line-horizontal.geojson
@@ -1,6 +1,9 @@
   {
     "type": "Feature",
-    "properties": {},
+    "properties": {
+      "stroke": "#F00",
+      "stroke-width": 6
+    },
     "geometry": {
       "type": "LineString",
       "coordinates": [
