--- packages/turf-line-offset/test/in/linestring-feature.geojson
@@ -1,25 +1,0 @@
-{
-  "type": "Feature",
-  "properties": {},
-  "geometry": {
-    "type": "LineString",
-    "coordinates": [
-      [
-        20.390625,
-        62.2679226294176
-      ],
-      [
-        27.0703125,
-        56.36525013685606
-      ],
-      [
-        22.5,
-        53.54030739150022
-      ],
-      [
-        26.015625,
-        49.83798245308484
-      ]
-    ]
-  }
-}
